Removed adapter, recombined teaser_manifold.py into a single file.
