--- adapter/__init__.py
@@ -1,10 +1,0 @@
-"""
-LeJEPA Optimizer Adapter
-
-Provides a production-ready optimizer implementing LeJEPA's official training
-configuration with statistical regularization.
-"""
-
-from .lejepa_adapter import *
-
-__all__ = ['LeJEPAAdapter']
